# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: Vyapaar AI API Specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /webhooks/whatsapp:
    get:
      operationId: verifyWhatsAppWebhook
      tags: [webhooks]
      summary: Verify WhatsApp Webhook
      description: Handles challenge verification from the Meta Developer platform.
      parameters:
        - name: hub.mode
          in: query
          required: true
          schema:
            type: string
        - name: hub.verify_token
          in: query
          required: true
          schema:
            type: string
        - name: hub.challenge
          in: query
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Verification challenge string
          content:
            text/plain:
              schema:
                type: string

    post:
      operationId: receiveWhatsAppMessage
      tags: [webhooks]
      summary: Ingest WhatsApp Business Event
      description: Receives and processes message hooks from WhatsApp API.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
      responses:
        "200":
          description: Acknowledged successfully
          content:
            application/json:
              schema:
                type: object
  /webhooks/whatsapp/logs:
    get:
      operationId: getWhatsAppLogs
      tags: [webhooks]
      summary: Get Audited WhatsApp Conversations
      description: Retrieves transaction history of AI replies.
      parameters:
        - name: orgId
          in: query
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Logs history list
          content:
            application/json:
              schema:
                type: array
                items:
                  type: object
                  properties:
                    id:
                      type: integer
                    orgId:
                      type: integer
                    senderPhone:
                      type: string
                    message:
                      type: string
                    reply:
                      type: string
                    intentDetected:
                      type: string
                    createdAt:
                      type: string
                  required: [id, orgId, senderPhone, message, reply, intentDetected]
  /webhooks/whatsapp/settings:
    get:
      operationId: getWhatsAppSettings
      tags: [webhooks]
      summary: Get WhatsApp Branch Auto-Reply Settings
      description: Retrieves branch toggle configurations.
      parameters:
        - name: branchId
          in: query
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Settings details
          content:
            application/json:
              schema:
                type: object
                properties:
                  branchId:
                    type: integer
                  autoReply:
                    type: boolean
                required: [branchId, autoReply]
    post:
      operationId: updateWhatsAppSettings
      tags: [webhooks]
      summary: Toggle WhatsApp Auto-Reply
      description: Saves active/inactive settings.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                branchId:
                  type: integer
                autoReply:
                  type: boolean
              required: [branchId, autoReply]
      responses:
        "200":
          description: Settings successfully updated
          content:
            application/json:
              schema:
                type: object
                properties:
                  branchId:
                    type: integer
                  autoReply:
                    type: boolean
                required: [branchId, autoReply]
...